% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{creelr}
\alias{creelr}
\alias{creelr-package}
\title{Analysis of Angler Creel Data}
\description{
Implements the Traditional Access Design for One Access Site.
}
\examples{
\dontrun{
 demo(toa)
}
}
\references{
Pollock, Kenneth Hugh, Cynthia M. Jones, and T. L. Brown. Angler Survey Methods and Their Applications in Fisheries Management. American Fisheries Society Special Publication 25. Bethesda, Md: American Fisheries Society, 1994.
}

