% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trad-one-access.R
\name{day_type}
\alias{day_type}
\title{Get Day Type}
\usage{
day_type(x, weekend = c("Saturday", "Sunday"), holidays = NULL)
}
\arguments{
\item{x}{A Date vector to get the day types for.}

\item{weekend}{A string indicating the days to be considered weekend.}

\item{holidays}{A Date containing holidays that can be treated as weekends.}
}
\value{
A factor of the day types ('Week' or 'Weekend').
}
\description{
Get Day Type
}
\examples{
day_type(seq(as.Date("2000-12-01"), as.Date("2000-12-31"), by = "day"))
day_type(seq(as.Date("2000-12-01"), as.Date("2000-12-31"), by = "day"),
  holidays = as.Date("2000-12-25"))
}

