% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_creel_data}
\alias{plot_creel_data}
\title{Plot Creel Data}
\usage{
plot_creel_data(data, weekend = c("Saturday", "Sunday"), holidays = NULL,
  parameter = "Catch")
}
\arguments{
\item{data}{A data.frame containing Date, Period, RodHours and Catch columns.}

\item{weekend}{A string indicating the days to be considered weekend.}

\item{holidays}{A Date vector of weekday holidays that can be treated as weekends.}

\item{parameter}{A string of the parameter to plot ('Catch' or 'Effort').}
}
\value{
A \code{\link{ggplot}} object.
}
\description{
Plot Creel Data
}
\examples{
library(ggplot2)
data(toa_example)
plot_creel_data(toa_example) + ylab("Daily Catch")
}

