% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_creel_estimates}
\alias{plot_creel_estimates}
\title{Plot Creel Estimates}
\usage{
plot_creel_estimates(data, parameter = "Catch")
}
\arguments{
\item{data}{A data.frame with the columns Month, Parameter, Estimate and SE.}

\item{parameter}{A string of the parameter to filter by.}
}
\value{
A \code{\link{ggplot}} object.
}
\description{
Plot Creel Estimates
}
\examples{
library(ggplot2)
data(toa_example)
plot_creel_estimates(trad_one_access(toa_example)) + ylab("Catch")
}

