% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sum-creel-estimates.R
\name{sum_creel_estimates}
\alias{sum_creel_estimates}
\title{Sum Creel Estimates}
\usage{
sum_creel_estimates(data, by = "Year", alpha = 0.05)
}
\arguments{
\item{data}{A data.frame with the results from trad_one_access}

\item{by}{A flag indicating the time period by which aggregation is made}

\item{alpha}{Significance level for confidence intervals}
}
\value{
A data.frame with the aggregated results
}
\description{
Sum Creel Estimates
}
\examples{
data(toa_dummy)
toa_results <- trad_one_access(toa_dummy)
sum_creel_estimates(toa_results)
}

