% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toa_dummy}
\alias{toa_dummy}
\title{xx Data}
\format{A data.frame with 5 columns:
\describe{
  \item{Date}{xx}
  \item{DayeType}{xx}
  \item{Period}{xx}
  \item{RodHours}{xx}
  \item{Catch}{xx}
}}
\usage{
toa_dummy
}
\description{
xx Data
}
\keyword{datasets}

