% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toa_example}
\alias{toa_example}
\title{Fishery Data}
\format{A data.frame with 4 columns:
\describe{
  \item{Date}{Sample dates from February 2010}
  \item{Period}{Time of day sampled}
  \item{RodHours}{Hours of fishing effort}
  \item{Catch}{Total number of fish, for every specie, kept and released}
}}
\usage{
toa_example
}
\description{
Fishery Data
}
\keyword{datasets}

