% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trad-one-access.R
\name{trad_one_access}
\alias{trad_one_access}
\title{Title}
\usage{
trad_one_access(data, am = 0.5, holidays = NULL, weekend = c("Saturday",
  "Sunday"), alpha = 0.05)
}
\arguments{
\item{data}{A data.frame xx}

\item{am}{A flag indicating xx}
}
\value{
A data.frame with xx
}
\description{
Title
}
\examples{
data(toa_dummy)
trad_one_access(toa_dummy)
}

