% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{creelr}
\alias{creelr}
\alias{creelr-package}
\title{Analysis of Angler Creel Data}
\description{
Implements the Traditional Access Design for One Access Site.
}
\examples{
\dontrun{
 demo(toa)
}
}
\references{
xx
}

