% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_creel_estimates}
\alias{plot_creel_estimates}
\title{Plot Creel Estimates}
\usage{
plot_creel_estimates(data, parameter = "Catch", x = "Month")
}
\arguments{
\item{data}{A data.frame with the columns Parameter, Estimate, Lower and Upper
and the variable for the x-axis.}

\item{parameter}{A string of the parameter to plot.}

\item{x}{A string of the variable to plot on the x-axis.}
}
\value{
A \code{\link{ggplot}} object.
}
\description{
Plot Creel Estimates
}
\examples{
library(ggplot2)
data(toa_example)
plot_creel_estimates(trad_one_access(toa_example)) + ylab("Catch")
}

