% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trad-one-access.R
\name{trad_one_access}
\alias{trad_one_access}
\title{Traditional One Access Estimates}
\usage{
trad_one_access(data, am = 0.5, holidays = NULL, weekend = c("Saturday",
  "Sunday"), alpha = 0.05, weighted = FALSE)
}
\arguments{
\item{data}{A data.frame containing Date, DayType, Period, RodHours and Catch}

\item{am}{A flag indicating the selection probility for the AM period}

\item{holidays}{A Date vector containing holidays that can be treated as weekends}

\item{weekend}{A string vector indicating the days to be considered weekend}

\item{alpha}{The significance level desired for the confidence intervals}

\item{weighted}{A logical value that indicated whether the sampling coverage should weight by Period or not}
}
\value{
A data.frame with the total effort and catch estimates, confidence intervals, in-week and weekend days and sampling coverage
}
\description{
Traditional One Access Estimates
}
\examples{
data(toa_dummy)
trad_one_access(toa_dummy)
}

